pyproject.toml
[project]
name = "photo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloud-sql-python-connector[pg8000]>=1.18.2",
    "google-adk>=1.3.0",
    "pg8000>=1.31.2",
    "sqlalchemy>=2.0.41",
    "toolbox-core>=0.2.1",
]
